Refactor and make more metadata internal
