+++ misk-admin/api/misk-admin.api
@@ -204,9 +204,6 @@ public final class com/squareup/protos/test/parsing/Warehouse$Builder : com/squa
 public abstract interface annotation class misk/web/dashboard/AdminDashboard : java/lang/annotation/Annotation {
 }
 
-public abstract interface annotation class misk/web/dashboard/AdminDashboardAccess : java/lang/annotation/Annotation {
-}
-
 public final class misk/web/dashboard/AdminDashboardModule : misk/inject/KAbstractModule {
 	public fun <init> (Z)V
 	public fun <init> (ZLmisk/web/metadata/config/ConfigMetadataAction$ConfigTabMode;)V
@@ -219,12 +216,6 @@ public final class misk/web/dashboard/AdminDashboardTestingModule : misk/inject/
 
 public final class misk/web/dashboard/BaseDashboardModule : misk/inject/KAbstractModule {
 	public fun <init> (Z)V
-}
-
-public final class misk/web/dashboard/ConfigDashboardTabModule : misk/inject/KAbstractModule {
-	public fun <init> (Z)V
-	public fun <init> (ZLmisk/web/metadata/config/ConfigMetadataAction$ConfigTabMode;)V
-	public synthetic fun <init> (ZLmisk/web/metadata/config/ConfigMetadataAction$ConfigTabMode;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
 public final class misk/web/dashboard/DashboardHomeUrl : misk/web/dashboard/ValidWebEntry {
@@ -409,10 +400,6 @@ public final class misk/web/dashboard/DashboardTheme {
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class misk/web/dashboard/DatabaseDashboardTabModule : misk/inject/KAbstractModule {
-	public fun <init> (Z)V
-}
-
 public final class misk/web/dashboard/EnvironmentToColorLookup {
 	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V
 	public final fun component1 ()Ljava/lang/String;
@@ -467,7 +454,7 @@ public final class misk/web/dashboard/MiskWebTabIndexAction : misk/web/actions/W
 	public static final field Companion Lmisk/web/dashboard/MiskWebTabIndexAction$Companion;
 	public static final field PATH Ljava/lang/String;
 	public fun <init> (Ljava/util/List;)V
-	public final fun get (Ljava/lang/String;Ljava/lang/String;)Ljava/lang/String;
+	public final fun get (Ljava/lang/String;)Ljava/lang/String;
 }
 
 public final class misk/web/dashboard/MiskWebTabIndexAction$Companion {
@@ -503,10 +490,6 @@ public final class misk/web/dashboard/MiskWebTheme {
 
 public final class misk/web/dashboard/MiskWebTheme$Companion {
 	public final fun getDEFAULT_THEME ()Lmisk/web/dashboard/MiskWebTheme;
-}
-
-public final class misk/web/dashboard/WebActionsDashboardTabModule : misk/inject/KAbstractModule {
-	public fun <init> (Z)V
 }
 
 public class misk/web/dashboard/WebTab : misk/web/dashboard/ValidWebEntry {
@@ -686,19 +669,14 @@ public final class misk/web/metadata/all/AllMetadataTabAction : misk/web/actions
 public final class misk/web/metadata/all/AllMetadataTabAction$Companion {
 }
 
-public final class misk/web/metadata/config/ConfigMetadata : misk/web/metadata/Metadata {
-	public fun <init> (Ljava/util/Map;)V
-	public final fun component1 ()Ljava/util/Map;
-	public final fun copy (Ljava/util/Map;)Lmisk/web/metadata/config/ConfigMetadata;
-	public static synthetic fun copy$default (Lmisk/web/metadata/config/ConfigMetadata;Ljava/util/Map;ILjava/lang/Object;)Lmisk/web/metadata/config/ConfigMetadata;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getResources ()Ljava/util/Map;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
+public final class misk/web/metadata/config/ConfigDashboardTabModule : misk/inject/KAbstractModule {
+	public fun <init> (Z)V
+	public fun <init> (ZLmisk/web/metadata/config/ConfigMetadataAction$ConfigTabMode;)V
+	public synthetic fun <init> (ZLmisk/web/metadata/config/ConfigMetadataAction$ConfigTabMode;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
 public final class misk/web/metadata/config/ConfigMetadataAction : misk/web/actions/WebAction {
-	public fun <init> (Lcom/google/inject/Provider;)V
+	public fun <init> ()V
 	public final fun getAll ()Lmisk/web/metadata/config/ConfigMetadataAction$Response;
 }
 
@@ -722,32 +700,8 @@ public final class misk/web/metadata/config/ConfigMetadataAction$Response {
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class misk/web/metadata/config/ConfigMetadataProvider : misk/web/metadata/MetadataProvider {
-	public fun <init> ()V
-	public synthetic fun get ()Ljava/lang/Object;
-	public fun get ()Lmisk/web/metadata/config/ConfigMetadata;
-	public fun getId ()Ljava/lang/String;
-}
-
-public final class misk/web/metadata/database/DatabaseHibernateMetadata : misk/web/metadata/Metadata {
-	public fun <init> (Ljava/util/List;)V
-	public final fun component1 ()Ljava/util/List;
-	public final fun copy (Ljava/util/List;)Lmisk/web/metadata/database/DatabaseHibernateMetadata;
-	public static synthetic fun copy$default (Lmisk/web/metadata/database/DatabaseHibernateMetadata;Ljava/util/List;ILjava/lang/Object;)Lmisk/web/metadata/database/DatabaseHibernateMetadata;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getHibernate ()Ljava/util/List;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class misk/web/metadata/database/DatabaseHibernateMetadataProvider : misk/web/metadata/MetadataProvider {
-	public field metadata Ljava/util/List;
-	public fun <init> ()V
-	public synthetic fun get ()Ljava/lang/Object;
-	public fun get ()Lmisk/web/metadata/database/DatabaseHibernateMetadata;
-	public fun getId ()Ljava/lang/String;
-	public final fun getMetadata ()Ljava/util/List;
-	public final fun setMetadata (Ljava/util/List;)V
+public final class misk/web/metadata/database/DatabaseDashboardTabModule : misk/inject/KAbstractModule {
+	public fun <init> (Z)V
 }
 
 public abstract interface class misk/web/metadata/database/DatabaseQueryFunctionMetadata {
@@ -920,21 +874,8 @@ public final class misk/web/metadata/jvm/JvmMetadataAction$JvmRuntimeResponse$Co
 	public final fun create (Ljava/lang/management/RuntimeMXBean;)Lmisk/web/metadata/jvm/JvmMetadataAction$JvmRuntimeResponse;
 }
 
-public final class misk/web/metadata/webaction/WebActionMetadataAction : misk/web/actions/WebAction {
-	public fun <init> (Lcom/google/inject/Provider;)V
-	public final fun getAll ()Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;
-	public final fun getProvider ()Lcom/google/inject/Provider;
-}
-
-public final class misk/web/metadata/webaction/WebActionMetadataAction$Response {
-	public fun <init> (Ljava/util/List;)V
-	public final fun component1 ()Ljava/util/List;
-	public final fun copy (Ljava/util/List;)Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;
-	public static synthetic fun copy$default (Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;Ljava/util/List;ILjava/lang/Object;)Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getWebActionMetadata ()Ljava/util/List;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
+public final class misk/web/metadata/webaction/WebActionsDashboardTabModule : misk/inject/KAbstractModule {
+	public fun <init> (Z)V
 }
 
 public final class misk/web/v2/DashboardHotwireTabAction : misk/web/actions/WebAction {
