+++ misk-admin/src/main/kotlin/misk/web/dashboard/MiskWebTabIndexAction.kt
@@ -28,7 +28,7 @@ class MiskWebTabIndexAction @Inject constructor(
   @Get("$PATH/{slug}/{rest:.*}")
   @ResponseContentType(MediaTypes.TEXT_HTML)
   @AdminDashboardAccess
-  fun get(@PathParam slug: String?, @PathParam rest: String?): String {
+  fun get(@PathParam slug: String?): String {
     val dashboardTab = dashboardTabs.firstOrNull { slug == it.slug }
       ?: throw NotFoundException("No Misk-Web tab found for slug: $slug")
     // TODO remove this hack when new Web Actions tab lands and old ones removed, v1 and v2 are in the same web-actions tab
