+++ misk-admin/src/main/kotlin/misk/web/metadata/database/DatabaseDashboardTabModule.kt
@@ -1,11 +1,11 @@
-package misk.web.dashboard
+package misk.web.metadata.database
 
 import misk.inject.KAbstractModule
 import misk.security.authz.AccessAnnotationEntry
 import misk.web.WebActionModule
-import misk.web.metadata.database.DatabaseQueryMetadata
-import misk.web.metadata.database.DatabaseQueryMetadataAction
-import misk.web.metadata.database.NoAdminDashboardDatabaseAccess
+import misk.web.dashboard.AdminDashboard
+import misk.web.dashboard.AdminDashboardAccess
+import misk.web.dashboard.DashboardModule
 
 /**
  * Installs Database dashboard tab which allows querying the database from a UI form.
@@ -16,14 +16,16 @@ class DatabaseDashboardTabModule(private val isDevelopment: Boolean): KAbstractM
     newMultibinder<DatabaseQueryMetadata>()
     install(WebActionModule.create<DatabaseQueryMetadataAction>())
 
-    install(DashboardModule.createMiskWebTab<AdminDashboard, AdminDashboardAccess>(
-      isDevelopment = isDevelopment,
-      slug = "database",
-      urlPathPrefix = "/_admin/database/",
-      developmentWebProxyUrl = "http://localhost:3202/",
-      menuLabel = "Database",
-      menuCategory = "Container Admin"
-    ))
+    install(
+      DashboardModule.createMiskWebTab<AdminDashboard, AdminDashboardAccess>(
+        isDevelopment = isDevelopment,
+        slug = "database",
+        urlPathPrefix = "/_admin/database/",
+        developmentWebProxyUrl = "http://localhost:3202/",
+        menuLabel = "Database",
+        menuCategory = "Container Admin"
+      )
+    )
 
     // Default access that doesn't allow any queries for unconfigured DbEntities
     multibind<AccessAnnotationEntry>().toInstance(
