+++ misk/api/misk.api
@@ -2188,6 +2188,22 @@ public final class misk/web/metadata/webaction/WebActionMetadata$ParameterMetaDa
 	public fun toString ()Ljava/lang/String;
 }
 
+public final class misk/web/metadata/webaction/WebActionMetadataAction : misk/web/actions/WebAction {
+	public fun <init> ()V
+	public final fun getAll ()Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;
+}
+
+public final class misk/web/metadata/webaction/WebActionMetadataAction$Response {
+	public fun <init> (Ljava/util/List;)V
+	public final fun component1 ()Ljava/util/List;
+	public final fun copy (Ljava/util/List;)Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;
+	public static synthetic fun copy$default (Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;Ljava/util/List;ILjava/lang/Object;)Lmisk/web/metadata/webaction/WebActionMetadataAction$Response;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getWebActionMetadata ()Ljava/util/List;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
 public final class misk/web/metadata/webaction/WebActionsMetadata : misk/web/metadata/Metadata {
 	public fun <init> (Ljava/util/List;)V
 	public final fun component1 ()Ljava/util/List;
