+++ misk/src/main/kotlin/misk/web/metadata/webaction/WebActionMetadataAction.kt
@@ -11,9 +11,9 @@ import misk.web.dashboard.AdminDashboardAccess
 import misk.web.mediatype.MediaTypes
 
 @Singleton
-class WebActionMetadataAction @Inject constructor(
-  val provider: Provider<WebActionsMetadata>
-) : WebAction {
+class WebActionMetadataAction @Inject constructor() : WebAction {
+  @Inject private lateinit var provider: Provider<WebActionsMetadata>
+
   @Get("/api/v1/webaction/metadata")
   @RequestContentType(MediaTypes.APPLICATION_JSON)
   @ResponseContentType(MediaTypes.APPLICATION_JSON)
