+++ samples/exemplar/src/main/kotlin/com/squareup/exemplar/dashboard/frontend/GraphD3JsPage.kt
@@ -8,13 +8,13 @@ import misk.turbo.turbo_frame
 import misk.web.Get
 import misk.web.ResponseContentType
 import misk.web.actions.WebAction
-import misk.web.dashboard.AdminDashboardAccess
 import misk.web.dashboard.HtmlLayout
 import misk.web.mediatype.MediaTypes
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
 import kotlinx.html.script
 import kotlinx.html.unsafe
+import misk.web.dashboard.AdminDashboardAccess
 
 /**
  * Example page that shows usage of D3.js graph library
